Move questionnaire data into separate class
